refactor: rename MainCoordinator to TagCoordinator
#13
